order, orderItem

+++ source/user-base-auth/src/main/java/com/base/auth/controller/CustomerController.java
@@ -2,7 +2,6 @@ package com.base.auth.controller;
 
 import com.base.auth.constant.UserBaseConstant;
 import com.base.auth.dto.ApiMessageDto;
-import com.base.auth.dto.ApiResponse;
 import com.base.auth.dto.ErrorCode;
 import com.base.auth.dto.ResponseListDto;
 import com.base.auth.dto.customer.CustomerDto;
@@ -50,6 +49,15 @@ public class CustomerController extends ABasicController {
     private CartRepository cartRepository;
 
     @Autowired
+    private CartItemRepository cartItemRepository;
+
+    @Autowired
+    private OrderRepository orderRepository;
+
+    @Autowired
+    private OrderItemRepository orderItemRepository;
+
+    @Autowired
     private CustomerMapper customerMapper;
 
     @Autowired
@@ -57,8 +65,8 @@ public class CustomerController extends ABasicController {
 
     @PostMapping(value = "/create", produces = MediaType.APPLICATION_JSON_VALUE)
     @PreAuthorize("hasRole('CUS_C')")
-    public ApiResponse<String> createCustomer(@Valid @RequestBody CreateCustomerForm createCustomerForm, BindingResult bindingResult) {
-        ApiResponse<String> apiMessageDto = new ApiResponse<>();
+    public ApiMessageDto<String> createCustomer(@Valid @RequestBody CreateCustomerForm createCustomerForm, BindingResult bindingResult) {
+        ApiMessageDto<String> apiMessageDto = new ApiMessageDto<>();
 
         // check username exist
         Account account = accountRepository.findAccountByUsername(createCustomerForm.getUsername());
@@ -97,7 +105,7 @@ public class CustomerController extends ABasicController {
         }
 
         // get district
-        Nation district = nationRepository.findByIdAndType(createCustomerForm.getProvinceId(), UserBaseConstant.NATION_KIND_DISTRICT)
+        Nation district = nationRepository.findByIdAndType(createCustomerForm.getDistrictId(), UserBaseConstant.NATION_KIND_DISTRICT)
                 .orElse(null);
         if (district == null) {
             apiMessageDto.setResult(false);
@@ -106,7 +114,7 @@ public class CustomerController extends ABasicController {
         }
 
         // get commune
-        Nation commune = nationRepository.findByIdAndType(createCustomerForm.getProvinceId(), UserBaseConstant.NATION_KIND_COMMUNE)
+        Nation commune = nationRepository.findByIdAndType(createCustomerForm.getCommuneId(), UserBaseConstant.NATION_KIND_COMMUNE)
                 .orElse(null);
         if (commune == null) {
             apiMessageDto.setResult(false);
@@ -117,19 +125,18 @@ public class CustomerController extends ABasicController {
         // create customer
         Customer customer = customerMapper.fromCreateCustomer(createCustomerForm);
 
+        customer.setAccount(account);
+        customer.setProvince(province);
+        customer.setDistrict(district);
+        customer.setCommune(commune);
+
+        customerRepository.save(customer);
+
         // create cart
         Cart cart = new Cart();
         cart.setCode(StringUtils.generateUniqueCode(6, cartRepository));
         cart.setCustomer(customer);
         cartRepository.save(cart);
-
-        customer.setAccount(account);
-        customer.setProvince(province);
-        customer.setDistrict(district);
-        customer.setCommune(commune);
-        customer.setCart(cart);
-
-        customerRepository.save(customer);
 
         apiMessageDto.setMessage("Create customer success");
         return apiMessageDto;
@@ -137,8 +144,8 @@ public class CustomerController extends ABasicController {
 
     @PutMapping(value = "/update", produces = MediaType.APPLICATION_JSON_VALUE)
     @PreAuthorize("hasRole('CUS_U')")
-    public ApiResponse<String> updateCustomer(@Valid @RequestBody UpdateCustomerForm updateCustomerForm, BindingResult bindingResult) {
-        ApiResponse<String> apiMessageDto = new ApiResponse<>();
+    public ApiMessageDto<String> updateCustomer(@Valid @RequestBody UpdateCustomerForm updateCustomerForm, BindingResult bindingResult) {
+        ApiMessageDto<String> apiMessageDto = new ApiMessageDto<>();
         Customer customer = customerRepository.findById(updateCustomerForm.getId()).orElse(null);
         if (customer == null) {
             apiMessageDto.setResult(false);
@@ -187,10 +194,10 @@ public class CustomerController extends ABasicController {
 
     @GetMapping(value = "/get/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     @PreAuthorize("hasRole('CUS_V')")
-    public ApiResponse<CustomerDto> get(@PathVariable("id") Long id) {
+    public ApiMessageDto<CustomerDto> get(@PathVariable("id") Long id) {
         Customer customer = customerRepository.findById(id).orElse(null);
         CustomerDto customerDto = customerMapper.fromCustomerToDto(customer);
-        ApiResponse<CustomerDto> apiMessageDto = new ApiResponse<>();
+        ApiMessageDto<CustomerDto> apiMessageDto = new ApiMessageDto<>();
         apiMessageDto.setData(customerDto);
         apiMessageDto.setMessage("Get customer success");
         return apiMessageDto;
@@ -198,7 +205,7 @@ public class CustomerController extends ABasicController {
 
     @GetMapping(value = "/list", produces = MediaType.APPLICATION_JSON_VALUE)
     @PreAuthorize("hasRole('CUS_L')")
-    public ApiResponse<ResponseListDto<List<CustomerDto>>> get(CustomerCriteria criteria, Pageable pageable) {
+    public ApiMessageDto<ResponseListDto<List<CustomerDto>>> get(CustomerCriteria criteria, Pageable pageable) {
         Page<Customer> pageData = customerRepository.findAll(criteria.getSpecification(), pageable);
 
         ResponseListDto<List<CustomerDto>> responseListDto = new ResponseListDto<>();
@@ -211,7 +218,7 @@ public class CustomerController extends ABasicController {
         responseListDto.setTotalElements(pageData.getTotalElements());
         responseListDto.setTotalPages(pageData.getTotalPages());
 
-        ApiResponse<ResponseListDto<List<CustomerDto>>> apiMessageDto = new ApiResponse<>();
+        ApiMessageDto<ResponseListDto<List<CustomerDto>>> apiMessageDto = new ApiMessageDto<>();
         apiMessageDto.setData(responseListDto);
         apiMessageDto.setMessage("Get list customers success");
         return apiMessageDto;
@@ -219,8 +226,8 @@ public class CustomerController extends ABasicController {
 
     @DeleteMapping(value = "/delete/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     @PreAuthorize("hasRole('CUS_D')")
-    public ApiResponse<String> delete(@PathVariable("id") Long id) {
-        ApiResponse<String> apiMessageDto = new ApiResponse<>();
+    public ApiMessageDto<String> delete(@PathVariable("id") Long id) {
+        ApiMessageDto<String> apiMessageDto = new ApiMessageDto<>();
         Customer customer = customerRepository.findById(id).orElse(null);
         if (customer == null) {
             apiMessageDto.setResult(false);
@@ -228,7 +235,21 @@ public class CustomerController extends ABasicController {
             return apiMessageDto;
         }
 
-        accountRepository.deleteById(customer.getAccount().getId());
+        Cart cart = cartRepository.findByCustomerId(customer.getId()).orElse(null);
+        if (cart != null) {
+            cartItemRepository.deleteAll(cart.getCartItems());
+            cartRepository.delete(cart);
+        }
+
+        List<Order> orders = orderRepository.findByCustomerId(id);
+        for (Order order : orders) {
+            orderItemRepository.deleteAll(order.getOrderItems());
+            orderRepository.delete(order);
+        }
+
+        if (customer.getAccount() != null) {
+            accountRepository.deleteById(customer.getAccount().getId());
+        }
 
         customerRepository.deleteById(id);
